breaking: completely re-write feature flags
Completely overhauls the feature flags for this library.

Feature flags are now formatted with a dash `-` instead of an underscore.

No default features are selected.

The hyper feature is now versioned: `hyper0_14` (corresponding to hyper 0.14). This effectively brings in tokio 1.x support and drops support for 0.x.

The curl feature is now named `isahc0_9` (pending an upgrade to isahc 1.x).

The rustls version has been upgraded across the board to 0.20, and the `hyper0_14` client now has support for using rustls as the tls adaptor.

The `unstable-config` feature has been removed.

Doc updates have been made.